requirements.txt
Django==2.2.1
django-extensions==2.1.7
django-filter==2.1.0
django-jalali==3.1.0
django-location-field==2.1.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
jdatetime==3.2.0
Pillow==6.0.0
pygraphviz==1.5
PyJWT==1.7.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
Werkzeug==0.15.4
